move BenchmarkParameters to types.py
