+++ stockfish/types.py
@@ -1,4 +1,6 @@
 from enum import Enum
+from dataclasses import dataclass
+import os
 
 
 class Piece(Enum):
@@ -20,3 +22,33 @@ class Capture(Enum):
     DIRECT_CAPTURE = "direct capture"
     EN_PASSANT = "en passant"
     NO_CAPTURE = "no capture"
+
+
+@dataclass
+class BenchmarkParameters:
+    ttSize: int = 16
+    threads: int = 1
+    limit: int = 13
+    fenFile: str = "default"
+    limitType: str = "depth"
+    evalType: str = "mixed"
+
+    def __post_init__(self):
+        self.ttSize = self.ttSize if self.ttSize in range(1, 128001) else 16
+        self.threads = self.threads if self.threads in range(1, 513) else 1
+        self.limit = self.limit if self.limit in range(1, 10001) else 13
+        self.fenFile = (
+            self.fenFile
+            if self.fenFile.endswith(".fen") and os.path.isfile(self.fenFile)
+            else "default"
+        )
+        self.limitType = (
+            self.limitType
+            if self.limitType in ["depth", "perft", "nodes", "movetime"]
+            else "depth"
+        )
+        self.evalType = (
+            self.evalType
+            if self.evalType in ["mixed", "classical", "NNUE"]
+            else "mixed"
+        )
